Added sorting filtering and pagination using Specification for WatchList

+++ tv-series-core/src/main/java/rs/ac/ni/pmf/rwa/tvseries/core/model/TvSeries.java
@@ -16,7 +16,7 @@ public class TvSeries {
     Integer numberOfEpisodes;
 
     Double averageRating;
-    Double usersRating;
+    Integer usersRating;
 
     Integer episodesWatched;
 }
